+++ library/core/src/main/java/com/google/android/exoplayer2/source/SampleQueue.java
@@ -15,23 +15,25 @@
  */
 package com.google.android.exoplayer2.source;
 
-import androidx.annotation.IntDef;
+import android.os.Looper;
 import androidx.annotation.Nullable;
 import com.google.android.exoplayer2.C;
 import com.google.android.exoplayer2.Format;
 import com.google.android.exoplayer2.FormatHolder;
 import com.google.android.exoplayer2.decoder.DecoderInputBuffer;
+import com.google.android.exoplayer2.drm.DrmInitData;
+import com.google.android.exoplayer2.drm.DrmSession;
+import com.google.android.exoplayer2.drm.DrmSessionManager;
 import com.google.android.exoplayer2.extractor.ExtractorInput;
 import com.google.android.exoplayer2.extractor.TrackOutput;
 import com.google.android.exoplayer2.source.SampleMetadataQueue.SampleExtrasHolder;
 import com.google.android.exoplayer2.upstream.Allocation;
 import com.google.android.exoplayer2.upstream.Allocator;
+import com.google.android.exoplayer2.util.Assertions;
 import com.google.android.exoplayer2.util.ParsableByteArray;
+import com.google.android.exoplayer2.util.Util;
 import java.io.EOFException;
 import java.io.IOException;
-import java.lang.annotation.Documented;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
 import java.nio.ByteBuffer;
 
 /** A queue of media samples. */
@@ -51,36 +53,18 @@ public class SampleQueue implements TrackOutput {
 
   }
 
-  /** Values returned by {@link #peekNext()}. */
-  @Documented
-  @Retention(RetentionPolicy.SOURCE)
-  @IntDef(
-      value = {
-        PEEK_RESULT_NOTHING,
-        PEEK_RESULT_FORMAT,
-        PEEK_RESULT_BUFFER_CLEAR,
-        PEEK_RESULT_BUFFER_ENCRYPTED
-      })
-  @interface PeekResult {}
-
-  /** Nothing is available for reading. */
-  public static final int PEEK_RESULT_NOTHING = 0;
-  /** A format change is available for reading */
-  public static final int PEEK_RESULT_FORMAT = 1;
-  /** A clear buffer is available for reading. */
-  public static final int PEEK_RESULT_BUFFER_CLEAR = 2;
-  /** An encrypted buffer is available for reading. */
-  public static final int PEEK_RESULT_BUFFER_ENCRYPTED = 3;
-
   public static final int ADVANCE_FAILED = -1;
 
   private static final int INITIAL_SCRATCH_SIZE = 32;
 
   private final Allocator allocator;
+  private final DrmSessionManager<?> drmSessionManager;
+  private final boolean playClearSamplesWithoutKeys;
   private final int allocationLength;
   private final SampleMetadataQueue metadataQueue;
   private final SampleExtrasHolder extrasHolder;
   private final ParsableByteArray scratch;
+  private final FormatHolder scratchFormatHolder;
 
   // References into the linked list of allocations.
   private AllocationNode firstAllocationNode;
@@ -89,6 +73,7 @@ public class SampleQueue implements TrackOutput {
 
   // Accessed only by the consuming thread.
   private Format downstreamFormat;
+  @Nullable private DrmSession<?> currentSession;
 
   // Accessed only by the loading thread (or the consuming thread when there is no loading thread).
   private boolean pendingFormatAdjustment;
@@ -99,14 +84,23 @@ public class SampleQueue implements TrackOutput {
   private UpstreamFormatChangedListener upstreamFormatChangeListener;
 
   /**
+   * Creates a sample queue.
+   *
    * @param allocator An {@link Allocator} from which allocations for sample data can be obtained.
+   * @param drmSessionManager The {@link DrmSessionManager} to obtain {@link DrmSession DrmSessions}
+   *     from.
    */
-  public SampleQueue(Allocator allocator) {
+  public SampleQueue(Allocator allocator, DrmSessionManager<?> drmSessionManager) {
     this.allocator = allocator;
+    this.drmSessionManager = drmSessionManager;
+    playClearSamplesWithoutKeys =
+        (drmSessionManager.getFlags() & DrmSessionManager.FLAG_PLAY_CLEAR_SAMPLES_WITHOUT_KEYS)
+            != 0;
     allocationLength = allocator.getIndividualAllocationLength();
     metadataQueue = new SampleMetadataQueue();
     extrasHolder = new SampleExtrasHolder();
     scratch = new ParsableByteArray(INITIAL_SCRATCH_SIZE);
+    scratchFormatHolder = new FormatHolder();
     firstAllocationNode = new AllocationNode(0, allocationLength);
     readAllocationNode = firstAllocationNode;
     writeAllocationNode = firstAllocationNode;
@@ -122,7 +116,7 @@ public class SampleQueue implements TrackOutput {
   }
 
   /**
-   * Resets the output.
+   * Resets the output and releases any held DRM resources.
    *
    * @param resetUpstreamFormat Whether the upstream format should be cleared. If set to false,
    *     samples queued after the reset (and before a subsequent call to {@link #format(Format)})
@@ -198,10 +192,15 @@ public class SampleQueue implements TrackOutput {
   // Called by the consuming thread.
 
   /**
-   * Returns whether a sample is available to be read.
+   * Throws an error that's preventing data from being read. Does nothing if no such error exists.
+   *
+   * @throws IOException The underlying error.
    */
-  public boolean hasNextSample() {
-    return metadataQueue.hasNextSample();
+  public void maybeThrowError() throws IOException {
+    // TODO: Avoid throwing if the DRM error is not preventing a read operation.
+    if (currentSession != null && currentSession.getState() == DrmSession.STATE_ERROR) {
+      throw Assertions.checkNotNull(currentSession.getError());
+    }
   }
 
   /**
@@ -292,6 +291,18 @@ public class SampleQueue implements TrackOutput {
     discardDownstreamTo(metadataQueue.discardToRead());
   }
 
+  /** Calls {@link #discardToEnd()} and releases any held DRM resources. */
+  public void preRelease() {
+    discardToEnd();
+    releaseDrmResources();
+  }
+
+  /** Calls {@link #reset()} and releases any held DRM resources. */
+  public void release() {
+    reset();
+    releaseDrmResources();
+  }
+
   /**
    * Discards to the end of the queue. The read position is also advanced.
    */
@@ -338,19 +349,18 @@ public class SampleQueue implements TrackOutput {
   }
 
   /**
-   * Returns a {@link PeekResult} depending on what a following call to {@link #read
-   * read(formatHolder, decoderInputBuffer, formatRequired= false, allowOnlyClearBuffers= false,
-   * loadingFinished= false, decodeOnlyUntilUs= 0)} would result in.
-   */
-  @PeekResult
-  public int peekNext() {
-    return metadataQueue.peekNext(downstreamFormat);
-  }
-
-  /**
    * Attempts to read from the queue.
    *
-   * @param formatHolder A {@link FormatHolder} to populate in the case of reading a format.
+   * <p>{@link Format Formats} read from the this method may be associated to a {@link DrmSession}
+   * through {@link FormatHolder#drmSession}, which is populated in two scenarios:
+   *
+   * <ul>
+   *   <li>The sample has a {@link Format} with a non-null {@link Format#drmInitData}.
+   *   <li>The {@link DrmSessionManager} is configured to use secure decoders for clear samples. See
+   *       {@link DrmSessionManager#FLAG_PLAY_CLEAR_SAMPLES_WITHOUT_KEYS}.
+   * </ul>
+   *
+   * @param outputFormatHolder A {@link FormatHolder} to populate in the case of reading a format.
    * @param buffer A {@link DecoderInputBuffer} to populate in the case of reading a sample or the
    *     end of the stream. If the end of the stream has been reached, the {@link
    *     C#BUFFER_FLAG_END_OF_STREAM} flag will be set on the buffer. If a {@link
@@ -359,33 +369,57 @@ public class SampleQueue implements TrackOutput {
    * @param formatRequired Whether the caller requires that the format of the stream be read even if
    *     it's not changing. A sample will never be read if set to true, however it is still possible
    *     for the end of stream or nothing to be read.
-   * @param allowOnlyClearBuffers If set to true, this method will not return encrypted buffers,
-   *     returning {@link C#RESULT_NOTHING_READ} (without advancing the read position) instead.
    * @param loadingFinished True if an empty queue should be considered the end of the stream.
    * @param decodeOnlyUntilUs If a buffer is read, the {@link C#BUFFER_FLAG_DECODE_ONLY} flag will
    *     be set if the buffer's timestamp is less than this value.
    * @return The result, which can be {@link C#RESULT_NOTHING_READ}, {@link C#RESULT_FORMAT_READ} or
    *     {@link C#RESULT_BUFFER_READ}.
    */
+  @SuppressWarnings("ReferenceEquality")
   public int read(
-      FormatHolder formatHolder,
+      FormatHolder outputFormatHolder,
       DecoderInputBuffer buffer,
       boolean formatRequired,
-      boolean allowOnlyClearBuffers,
       boolean loadingFinished,
       long decodeOnlyUntilUs) {
+
+    boolean readFlagFormatRequired = false;
+    boolean readFlagAllowOnlyClearBuffers = false;
+    boolean onlyPropagateFormatChanges = false;
+
+    if (downstreamFormat == null || formatRequired) {
+      readFlagFormatRequired = true;
+    } else if (drmSessionManager != DrmSessionManager.DUMMY
+        && downstreamFormat.drmInitData != null
+        && Assertions.checkNotNull(currentSession).getState()
+            != DrmSession.STATE_OPENED_WITH_KEYS) {
+      if (playClearSamplesWithoutKeys) {
+        // Content is encrypted and keys are not available, but clear samples are ok for reading.
+        readFlagAllowOnlyClearBuffers = true;
+      } else {
+        // We must not read any samples, but we may still read a format or the end of stream.
+        // However, because the formatRequired argument is false, we should not propagate a read
+        // format unless it is different than the current format.
+        onlyPropagateFormatChanges = true;
+        readFlagFormatRequired = true;
+      }
+    }
+
     int result =
         metadataQueue.read(
-            formatHolder,
+            scratchFormatHolder,
             buffer,
-            formatRequired,
-            allowOnlyClearBuffers,
+            readFlagFormatRequired,
+            readFlagAllowOnlyClearBuffers,
             loadingFinished,
             downstreamFormat,
             extrasHolder);
     switch (result) {
       case C.RESULT_FORMAT_READ:
-        downstreamFormat = formatHolder.format;
+        if (onlyPropagateFormatChanges && downstreamFormat == scratchFormatHolder.format) {
+          return C.RESULT_NOTHING_READ;
+        }
+        onFormat(Assertions.checkNotNull(scratchFormatHolder.format), outputFormatHolder);
         return C.RESULT_FORMAT_READ;
       case C.RESULT_BUFFER_READ:
         if (!buffer.isEndOfStream()) {
@@ -399,6 +433,35 @@ public class SampleQueue implements TrackOutput {
         return C.RESULT_BUFFER_READ;
       case C.RESULT_NOTHING_READ:
         return C.RESULT_NOTHING_READ;
+      default:
+        throw new IllegalStateException();
+    }
+  }
+
+  /**
+   * Returns whether there is data available for reading.
+   *
+   * <p>Note: If the stream has ended then a buffer with the end of stream flag can always be read
+   * from {@link #read}. Hence an ended stream is always ready.
+   *
+   * @param loadingFinished Whether no more samples will be written to the sample queue. When true,
+   *     this method returns true if the sample queue is empty, because an empty sample queue means
+   *     the end of stream has been reached. When false, this method returns false if the sample
+   *     queue is empty.
+   */
+  public boolean isReady(boolean loadingFinished) {
+    @SampleMetadataQueue.PeekResult int nextInQueue = metadataQueue.peekNext(downstreamFormat);
+    switch (nextInQueue) {
+      case SampleMetadataQueue.PEEK_RESULT_NOTHING:
+        return loadingFinished;
+      case SampleMetadataQueue.PEEK_RESULT_FORMAT:
+        return true;
+      case SampleMetadataQueue.PEEK_RESULT_BUFFER_CLEAR:
+        return currentSession == null || playClearSamplesWithoutKeys;
+      case SampleMetadataQueue.PEEK_RESULT_BUFFER_ENCRYPTED:
+        return drmSessionManager == DrmSessionManager.DUMMY
+            || Assertions.checkNotNull(currentSession).getState()
+                == DrmSession.STATE_OPENED_WITH_KEYS;
       default:
         throw new IllegalStateException();
     }
@@ -748,9 +811,55 @@ public class SampleQueue implements TrackOutput {
     return format;
   }
 
+  /** Releases any held DRM resources. */
+  private void releaseDrmResources() {
+    if (currentSession != null) {
+      currentSession.releaseReference();
+      currentSession = null;
+    }
+  }
+
   /**
-   * A node in a linked list of {@link Allocation}s held by the output.
+   * Updates the current format and manages any necessary DRM resources.
+   *
+   * @param format The format read from upstream.
+   * @param outputFormatHolder The output {@link FormatHolder}.
    */
+  private void onFormat(Format format, FormatHolder outputFormatHolder) {
+    outputFormatHolder.format = format;
+    boolean isFirstFormat = downstreamFormat == null;
+    DrmInitData oldDrmInitData = isFirstFormat ? null : downstreamFormat.drmInitData;
+    downstreamFormat = format;
+    if (drmSessionManager == DrmSessionManager.DUMMY) {
+      // Avoid attempting to acquire a session using the dummy DRM session manager. It's likely that
+      // the media source creation has not yet been migrated and the renderer can acquire the
+      // session for the read DRM init data.
+      // TODO: Remove once renderers are migrated [Internal ref: b/122519809].
+      return;
+    }
+    outputFormatHolder.includesDrmSession = true;
+    outputFormatHolder.drmSession = currentSession;
+    if (!isFirstFormat && Util.areEqual(oldDrmInitData, format.drmInitData)) {
+      // Nothing to do.
+      return;
+    }
+    // Ensure we acquire the new session before releasing the previous one in case the same session
+    // can be used for both DrmInitData.
+    DrmSession<?> previousSession = currentSession;
+    DrmInitData drmInitData = downstreamFormat.drmInitData;
+    Looper playbackLooper = Assertions.checkNotNull(Looper.myLooper());
+    currentSession =
+        drmInitData != null
+            ? drmSessionManager.acquireSession(playbackLooper, drmInitData)
+            : drmSessionManager.acquirePlaceholderSession(playbackLooper);
+    outputFormatHolder.drmSession = currentSession;
+
+    if (previousSession != null) {
+      previousSession.releaseReference();
+    }
+  }
+
+  /** A node in a linked list of {@link Allocation}s held by the output. */
   private static final class AllocationNode {
 
     /**