+++ library/hls/src/main/java/com/google/android/exoplayer2/source/hls/playlist/HlsPlaylistParser.java
@@ -48,10 +48,12 @@ import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.NoSuchElementException;
 import java.util.Queue;
 import java.util.TreeMap;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
+import org.checkerframework.checker.nullness.qual.EnsuresNonNullIf;
 import org.checkerframework.checker.nullness.qual.PolyNull;
 
 /**
@@ -338,6 +340,9 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
             parseOptionalStringAttr(line, REGEX_SUBTITLES, variableDefinitions);
         String closedCaptionsGroupId =
             parseOptionalStringAttr(line, REGEX_CLOSED_CAPTIONS, variableDefinitions);
+        if (!iterator.hasNext()) {
+          throw new ParserException("#EXT-X-STREAM-INF tag must be followed by another line");
+        }
         line =
             replaceVariableReferences(
                 iterator.next(), variableDefinitions); // #EXT-X-STREAM-INF's URI.
@@ -381,7 +386,9 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
         Assertions.checkState(variant.format.metadata == null);
         HlsTrackMetadataEntry hlsMetadataEntry =
             new HlsTrackMetadataEntry(
-                /* groupId= */ null, /* name= */ null, urlToVariantInfos.get(variant.url));
+                /* groupId= */ null,
+                /* name= */ null,
+                Assertions.checkNotNull(urlToVariantInfos.get(variant.url)));
         deduplicatedVariants.add(
             variant.copyWithFormat(
                 variant.format.copyWithMetadata(new Metadata(hlsMetadataEntry))));
@@ -536,6 +543,7 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
         sessionKeyDrmInitData);
   }
 
+  @Nullable
   private static Variant getVariantWithAudioGroup(ArrayList<Variant> variants, String groupId) {
     for (int i = 0; i < variants.size(); i++) {
       Variant variant = variants.get(i);
@@ -546,6 +554,7 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
     return null;
   }
 
+  @Nullable
   private static Variant getVariantWithVideoGroup(ArrayList<Variant> variants, String groupId) {
     for (int i = 0; i < variants.size(); i++) {
       Variant variant = variants.get(i);
@@ -931,19 +940,20 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
     private final BufferedReader reader;
     private final Queue<String> extraLines;
 
-    private String next;
+    @Nullable private String next;
 
     public LineIterator(Queue<String> extraLines, BufferedReader reader) {
       this.extraLines = extraLines;
       this.reader = reader;
     }
 
+    @EnsuresNonNullIf(expression = "next", result = true)
     public boolean hasNext() throws IOException {
       if (next != null) {
         return true;
       }
       if (!extraLines.isEmpty()) {
-        next = extraLines.poll();
+        next = Assertions.checkNotNull(extraLines.poll());
         return true;
       }
       while ((next = reader.readLine()) != null) {
@@ -955,13 +965,15 @@ public final class HlsPlaylistParser implements ParsingLoadable.Parser<HlsPlayli
       return false;
     }
 
+    /** Return the next line, or throw {@link NoSuchElementException} if none. */
     public String next() throws IOException {
-      String result = null;
       if (hasNext()) {
-        result = next;
+        String result = next;
         next = null;
+        return result;
+      } else {
+        throw new NoSuchElementException();
       }
-      return result;
     }
 
   }